requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
docutils==0.16
imagesize==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pygments==2.6.1
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
discord.py==1.4.1
flake8==3.8.3
idna==2.9
importlib-metadata==1.6.1
mccabe==0.6.1
multidict==4.7.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.12.0
PyYAML==5.3.1
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.0.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tabulate==0.8.7
urllib3==1.25.9
validators==0.14.2
websockets==8.1
yarl==1.4.2
zipp==3.1.0
